add auth skeleton

+++ metacat-common-server/src/main/java/com/netflix/metacat/common/server/connectors/AuthEnabledConnectorPartitionService.java
@@ -1,16 +1,16 @@
 package com.netflix.metacat.common.server.connectors;
 
+import com.netflix.metacat.common.MetacatRequestContext;
 import com.netflix.metacat.common.QualifiedName;
 import com.netflix.metacat.common.dto.Pageable;
 import com.netflix.metacat.common.dto.Sort;
 import com.netflix.metacat.common.exception.MetacatTooManyRequestsException;
+import com.netflix.metacat.common.exception.MetacatUnAuthorizedException;
+import com.netflix.metacat.common.server.api.authorization.Authorization;
+import com.netflix.metacat.common.server.api.authorization.AuthorizationStatus;
 import com.netflix.metacat.common.server.api.ratelimiter.RateLimiter;
 import com.netflix.metacat.common.server.api.ratelimiter.RateLimiterRequestContext;
-import com.netflix.metacat.common.server.connectors.model.PartitionInfo;
-import com.netflix.metacat.common.server.connectors.model.PartitionListRequest;
-import com.netflix.metacat.common.server.connectors.model.PartitionsSaveRequest;
-import com.netflix.metacat.common.server.connectors.model.PartitionsSaveResponse;
-import com.netflix.metacat.common.server.connectors.model.TableInfo;
+import com.netflix.metacat.common.server.connectors.model.*;
 import com.netflix.metacat.common.server.util.MetacatContextManager;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import lombok.Getter;
@@ -23,25 +23,23 @@ import java.util.List;
 import java.util.Map;
 
 /**
- * Connector that throttles calls to the connector based on the contextual request name
- * and the resource. Not all APIs can be throttled since we may not have a resource
- * but those are a small minority
+ * Connector that authorizes requests based on the request context.
  */
 @Slf4j
 @RequiredArgsConstructor
-public class ThrottlingConnectorPartitionService implements ConnectorPartitionService {
+public class AuthEnabledConnectorPartitionService implements ConnectorPartitionService {
     @Getter
     @NonNull
     private final ConnectorPartitionService delegate;
     @NonNull
-    private final RateLimiter rateLimiter;
+    private final Authorization authorization;
 
     @Override
     public List<PartitionInfo> getPartitions(final ConnectorRequestContext context,
                                              final QualifiedName table,
                                              final PartitionListRequest partitionsRequest,
                                              final TableInfo tableInfo) {
-        checkThrottling(MetacatContextManager.getContext().getRequestName(), table);
+        authorize(MetacatContextManager.getContext(), table.toString());
         return delegate.getPartitions(context, table, partitionsRequest, tableInfo);
     }
 
@@ -49,7 +47,7 @@ public class ThrottlingConnectorPartitionService implements ConnectorPartitionSe
     public PartitionsSaveResponse savePartitions(final ConnectorRequestContext context,
                                                  final QualifiedName table,
                                                  final PartitionsSaveRequest partitionsSaveRequest) {
-        checkThrottling(MetacatContextManager.getContext().getRequestName(), table);
+        authorize(MetacatContextManager.getContext(), table.toString());
         return delegate.savePartitions(context, table, partitionsSaveRequest);
     }
 
@@ -58,7 +56,7 @@ public class ThrottlingConnectorPartitionService implements ConnectorPartitionSe
                                  final QualifiedName tableName,
                                  final List<String> partitionNames,
                                  final TableInfo tableInfo) {
-        checkThrottling(MetacatContextManager.getContext().getRequestName(), tableName);
+        authorize(MetacatContextManager.getContext(), tableName.toString());
         delegate.deletePartitions(context, tableName, partitionNames, tableInfo);
     }
 
@@ -66,7 +64,7 @@ public class ThrottlingConnectorPartitionService implements ConnectorPartitionSe
     public int getPartitionCount(final ConnectorRequestContext context,
                                  final QualifiedName table,
                                  final TableInfo tableInfo) {
-        checkThrottling(MetacatContextManager.getContext().getRequestName(), table);
+        authorize(MetacatContextManager.getContext(), table.toString());
         return delegate.getPartitionCount(context, table, tableInfo);
     }
 
@@ -74,6 +72,7 @@ public class ThrottlingConnectorPartitionService implements ConnectorPartitionSe
     public Map<String, List<QualifiedName>> getPartitionNames(final ConnectorRequestContext context,
                                                               final List<String> uris,
                                                               final boolean prefixSearch) {
+        authorize(MetacatContextManager.getContext(), "N/A");
         return delegate.getPartitionNames(context, uris, prefixSearch);
     }
 
@@ -82,7 +81,7 @@ public class ThrottlingConnectorPartitionService implements ConnectorPartitionSe
                                          final QualifiedName table,
                                          final PartitionListRequest partitionsRequest,
                                          final TableInfo tableInfo) {
-        checkThrottling(MetacatContextManager.getContext().getRequestName(), table);
+        authorize(MetacatContextManager.getContext(), table.toString());
         return delegate.getPartitionKeys(context, table, partitionsRequest, tableInfo);
     }
 
@@ -91,38 +90,38 @@ public class ThrottlingConnectorPartitionService implements ConnectorPartitionSe
                                          final QualifiedName table,
                                          final PartitionListRequest partitionsRequest,
                                          final TableInfo tableInfo) {
-        checkThrottling(MetacatContextManager.getContext().getRequestName(), table);
+        authorize(MetacatContextManager.getContext(), table.toString());
         return delegate.getPartitionUris(context, table, partitionsRequest, tableInfo);
     }
 
     @Override
     public void create(final ConnectorRequestContext context, final PartitionInfo resource) {
-        checkThrottling(MetacatContextManager.getContext().getRequestName(), resource.getName());
+        authorize(MetacatContextManager.getContext(), resource.getName().toString());
         delegate.create(context, resource);
     }
 
     @Override
     public void update(final ConnectorRequestContext context, final PartitionInfo resource) {
-        checkThrottling(MetacatContextManager.getContext().getRequestName(), resource.getName());
+        authorize(MetacatContextManager.getContext(), resource.getName().toString());
         delegate.update(context, resource);
     }
 
     @Override
     public void delete(final ConnectorRequestContext context, final QualifiedName name) {
-        checkThrottling(MetacatContextManager.getContext().getRequestName(), name);
+        authorize(MetacatContextManager.getContext(), name.toString());
         delegate.delete(context, name);
     }
 
     @Override
     public PartitionInfo get(final ConnectorRequestContext context, final QualifiedName name) {
-        checkThrottling(MetacatContextManager.getContext().getRequestName(), name);
+        authorize(MetacatContextManager.getContext(), name.toString());
         return delegate.get(context, name);
     }
 
     @Override
     @SuppressFBWarnings
     public boolean exists(final ConnectorRequestContext context, final QualifiedName name) {
-        checkThrottling(MetacatContextManager.getContext().getRequestName(), name);
+        authorize(MetacatContextManager.getContext(), name.toString());
         return delegate.exists(context, name);
     }
 
@@ -130,7 +129,7 @@ public class ThrottlingConnectorPartitionService implements ConnectorPartitionSe
     public List<PartitionInfo> list(final ConnectorRequestContext context, final QualifiedName name,
                                     @Nullable final QualifiedName prefix, @Nullable final Sort sort,
                                     @Nullable final Pageable pageable) {
-        checkThrottling(MetacatContextManager.getContext().getRequestName(), name);
+        authorize(MetacatContextManager.getContext(), name.toString());
         return delegate.list(context, name, prefix, sort, pageable);
     }
 
@@ -138,23 +137,24 @@ public class ThrottlingConnectorPartitionService implements ConnectorPartitionSe
     public List<QualifiedName> listNames(final ConnectorRequestContext context, final QualifiedName name,
                                          @Nullable final QualifiedName prefix,
                                          @Nullable final Sort sort, @Nullable final Pageable pageable) {
-        checkThrottling(MetacatContextManager.getContext().getRequestName(), name);
+        authorize(MetacatContextManager.getContext(), name.toString());
         return delegate.listNames(context, name, prefix, sort, pageable);
     }
 
     @Override
     public void rename(final ConnectorRequestContext context, final QualifiedName oldName,
                        final QualifiedName newName) {
-        checkThrottling(MetacatContextManager.getContext().getRequestName(), oldName);
+        authorize(MetacatContextManager.getContext(), oldName.toString());
         delegate.rename(context, oldName, newName);
     }
 
-    private void checkThrottling(final String requestName, final QualifiedName resource) {
-        if (rateLimiter.hasExceededRequestLimit(new RateLimiterRequestContext(requestName, resource))) {
-            final String errorMsg = String.format("Too many requests for resource %s. Request: %s",
-                resource, requestName);
+    private void authorize(final MetacatRequestContext context, final String resource) {
+        final AuthorizationStatus status = authorization.isAuthorized(context);
+        if (!status.isAuthorized()) {
+            final String errorMsg = String.format("Forbidden request %s for resource %s. Details: %s",
+                MetacatContextManager.getContext().getRequestName(), resource, status.getDetails());
             log.warn(errorMsg);
-            throw new MetacatTooManyRequestsException(errorMsg);
+            throw new MetacatUnAuthorizedException(errorMsg);
         }
     }
